Add Prisma setup and User model
Prisma has been added as a new ORM, and a user model has been created in a Prisma schema file. This was done to achieve a more robust and reliable data access solution. Prisma client and Prisma packages have also been added to the package.json file to ensure that they are installed with other project dependencies.