Switch application to use SQLite and make Docker setup more robust
Deleted the .dockerignore and Prisma's migration.sql file due to the change in the database from PostgreSQL to SQLite. Updated the package-lock.json with the latest versions of dependencies. The prisma config file was tweaked to use SQLite and environmental variables for the database URL. Docker-compose was updated to include the 'prisma' directory in the shared volumes. Dockerfile now handles improved user permissions for more robust Docker setup. It now creates a new user and sets directory permissions for improved security. Also, the database switch requires automatic Prettier corrections to be applied, files to be built again, and changed start.sh permissions to be executable.