Remove unused 'hello.api.ts' file
The 'hello.api.ts' file was removed because it was unused. This API route was originally created for introductory and instructional purposes. After an understanding was established of how API routes work in Next.js, this file no longer served a purpose in the project. This deletion helps keep the project clean and free of unnecessary bloat.