Add axios API call in register page and remove email from form schema
In this commit, we made improvements to the registration page. We added the axios API call to do a POST request to '/users' when handling registration. This change allows the form data to be sent to the server. Additionally, we removed the 'email' field from the registration form schema as it is no longer needed