Add check for existing username before registration
This commit introduces a validation for username uniqueness when a new user is trying to register. If a user with the entered username already exists, the response will contain a corresponding error message. This error message is then handled on client-side registration page to alert the user about this validation case. This change improves user experience by providing clear feedback about the issue and prevents potential errors in case of duplicated usernames.