"Add time intervals selection in registration process
A new page was created for the registration process that adds the functionality to select available time intervals for each day of the week. This was achieved through creating a new 'time-intervals' index.page.tsx and styles.ts files. Additionally, a session function was added to the [...nextauth].api.ts file to ensure user sessions are considered when selecting these time intervals. This change allows users to define their availability during the registration process, providing a more accurate and detailed user profile."