Implement data posting for user time intervals
This commit enables user time's intervals data to be posted to the backend API. Previously, this data was console logged. The function 'handleSetTimeIntervals' has been modified to post data to the 'users/time-intervals' endpoint. Additionally, a new API handler for 'users/time-intervals' has been implemented to accept the POST request.